Add macros for "et al.", "etc.", "i.e.", and "e.g."
The macros handle spacing after periods.  For "et al." and "etc.", the
macros are defined such that a period following the macro in source
code is succeeded by inter-sentence spacing whereas any non-period
character after the macro results in the addition of a period (to
finish the Latin abbreviation) and an inter-word space.

The macros for "i.e." and "e.g." simply insert an inter-word space
after the Latin.